+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+ - Holding the camera home button will store the camera home position
+
 ## [11.2.0] - 2020-02-20
 ### Added
  - The center D-Pad button returns the selected camera to the home position